Problem: latest parity/scale-codec failed to compile (CRO-161) #120 (#133)
Solution: converted usize fields (in datatypes that use scale-codec) to u64

+++ chain-core/src/state/mod.rs
@@ -66,7 +66,7 @@ pub struct CouncilNode {
     // Tendermint consensus validator-associated public key
     pub consensus_pubkey: TendermintValidatorPubKey,
     // update counter
-    pub nonce: usize,
+    pub nonce: u64,
     // TODO: public keys / addresses for other operations
 }
